Removed `testing.go` file
This is not what I originally thought it was. This is to export useful
testing helpers/types to help a library users to write tests.

Also, the `init()` function of this will run all the time (e.g.
when the production server start) which is *not* what we want.
